add some elements
