+++ frontend/app/_utils/formatLocation.ts
@@ -1,0 +1,20 @@
+export default function SchoolLocation({management}){   
+   
+   const address = management?.schoolLocation?.[0]?.location;
+    const mapsUrl = address
+        ? `https://www.google.com/maps/search/?api=1&query=${encodeURIComponent(address)}`
+        : null;
+    const { displayAddress, regionLine } = address
+        ? (() => {
+            const parts = address.split(",").map(s => s.trim());
+            const [name, houseNumber, street, city, region, , province, , country] = parts;
+            const displayAddress = [name, [houseNumber, street].filter(Boolean).join(" ")]
+                .filter(Boolean)
+                .join(", ");
+            const regionLine = [city, region, province, country].filter(Boolean).join(", ");
+            return { displayAddress, regionLine };
+        })()
+        : { displayAddress: null, regionLine: null };
+
+    return [mapsUrl, displayAddress, regionLine];
+}
