+++ frontend/app/private/(authenticated)/styles.css
@@ -1,0 +1,3 @@
+body {
+    background-color: white;
+}
